requirements.txt
oracledb
bcrypt==4.0.1
python-dotenv==1.0.0
